refactor: move import utils to separate file
