+++ lib/rules/new-module-imports.js
@@ -2,7 +2,7 @@
 
 const MAPPING = require('ember-rfc176-data');
 const { buildMessage, getFullNames, isDestructuring } = require('../utils/new-module');
-const { getSourceModuleNameForIdentifier } = require('../utils/utils');
+const { getSourceModuleNameForIdentifier } = require('../utils/import');
 
 const GLOBALS = MAPPING.reduce((memo, exportDefinition) => {
   if (exportDefinition.deprecated) {
