+++ lib/rules/no-classic-classes.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const { getSourceModuleNameForIdentifier } = require('../utils/utils');
+const { getSourceModuleNameForIdentifier } = require('../utils/import');
 const { isObjectExpression } = require('../utils/types');
 
 const ERROR_MESSAGE_NO_CLASSIC_CLASSES =
