+++ lib/rules/require-tagless-components.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const { isEmberComponent } = require('../utils/ember');
-const { getSourceModuleNameForIdentifier } = require('../utils/utils');
+const { getSourceModuleNameForIdentifier } = require('../utils/import');
 const {
   isCallExpression,
   isIdentifier,
