+++ lib/rules/use-ember-data-rfc-395-imports.js
@@ -2,7 +2,7 @@
 
 const MAPPINGS = require('@ember-data/rfc395-data');
 const { buildFix, buildMessage, getFullNames, isDestructuring } = require('../utils/new-module');
-const { getSourceModuleNameForIdentifier } = require('../utils/utils');
+const { getSourceModuleNameForIdentifier } = require('../utils/import');
 
 /**
  * This function returns an object like this:
