+++ lib/utils/ember.js
@@ -1,6 +1,7 @@
 'use strict';
 
 const utils = require('./utils');
+const importUtils = require('./import');
 const types = require('./types');
 const assert = require('assert');
 
@@ -177,7 +178,10 @@ function isEmberCoreModule(context, node, moduleName) {
     if (!node.superClass) {
       return false;
     }
-    const superClassImportPath = utils.getSourceModuleNameForIdentifier(context, node.superClass);
+    const superClassImportPath = importUtils.getSourceModuleNameForIdentifier(
+      context,
+      node.superClass
+    );
 
     if (superClassImportPath === CORE_MODULE_IMPORT_PATHS[moduleName]) {
       return true;
