+++ lib/utils/utils.js
@@ -3,13 +3,11 @@
 const {
   isCallExpression,
   isIdentifier,
-  isImportDeclaration,
   isLiteral,
   isMemberExpression,
   isNewExpression,
   isObjectPattern,
 } = require('../utils/types');
-const assert = require('assert');
 
 module.exports = {
   collectObjectPatternBindings,
@@ -18,8 +16,6 @@ module.exports = {
   getParent,
   getPropertyValue,
   getSize,
-  getSourceModuleName,
-  getSourceModuleNameForIdentifier,
   isEmptyMethod,
   isGlobalCallExpression,
   parseArgs,
@@ -200,41 +196,6 @@ function collectObjectPatternBindings(node, initialObjToBinding) {
  */
 function isEmptyMethod(node) {
   return node.value.body && node.value.body.body && node.value.body.body.length <= 0;
-}
-
-/**
- * Gets the name of the module that an identifier was imported from,
- * if it was imported
- *
- * @param {Object} context the context of the ESLint rule
- * @param {node} node the Idenfifier to find an import for
- * @returns {string | undefined} The name of the module the identifier was imported from, if it was imported
- */
-function getSourceModuleNameForIdentifier(context, node) {
-  const [program] = context.getAncestors(node);
-  const importDeclaration = program.body
-    .filter(isImportDeclaration)
-    .find(importDeclaration =>
-      importDeclaration.specifiers.some(
-        specifier => specifier.local.name === getSourceModuleName(node)
-      )
-    );
-
-  return importDeclaration ? importDeclaration.source.value : undefined;
-}
-
-function getSourceModuleName(node) {
-  if (isMemberExpression(node) && node.object) {
-    return getSourceModuleName(node.object);
-  } else if (isIdentifier(node)) {
-    return node.name;
-  } else {
-    assert(
-      false,
-      '`getSourceModuleName` should only be called on `MemberExpression` or `Identifier`'
-    );
-    return undefined;
-  }
 }
 
 /**
